New: MEI 2015 Customizations
This commit moves the MEI 2015 changes from the incubator into the customizations directory. It moves the ODD file and the testing file to their proper places, but has not moved the changes to the Guidelines. This should be handled in a separate pull request.

This pull request requires #237 to be merged to build correctly.