Added two sections about new generic ornament element to the guidelines
This commit adds two new sections about the new "ornament" element to two chapters of the guidelines.
The first section "Ornaments" is added to the chapter describing shared elements (currently chap 1).
The second section is added to the chapter describing the encoding of ornamentation for CMN (currently chap 8). This also includes an example and a new image.

Refs #253