Fixed: Allow @visible on <accid>
This commit adds the @visible parameter to <accid>, providing a way to record an accidental that is present, but hidden. Note that this is not to be confused with the function of cautionary accidentals (i.e., accidentals that would normally be shown but for the convention that they were already shown in the bar.

Also included in this commit is a test case to check that <accid> takes @visible.