New: Build script now checks schematron rules as well
When the schemas are built, the schematron rules are extracted and transformed into an XSLT that is stored in the build/schematron directory

When the tests are run, saxon is used to apply the validation XSLT to the MEI files. A custom validation checker xquery file is then used to extract any errors from the schematron validation file and displays them in the console.

Fixes #264
Refs #262