Add model.chordPart
This is a new class for things that can be inside a chord. It replaces <note> and <artic> as <chord> content with a reference to this model class and places <note> and <artic> in the new class. This implementation of <chord> contents is better for customisations that need to add contents to <chord>.

Refs #298