Added model.rdgPart.text to content model of <lem> and <rdg>
This allows textual elements of scores which are members of that model to be the subject of critical of apparatus. Currently, as well as several more generic models, the content model of <lem> and <rdg> included only model.rdgPart.critapp allowing musical content. But no sufficiently specific model allowed textual elements of scores to be in critical apparatus.