Improved conversion to PureODD; generated and manually adjusted new PureODD source.
This commit includes improvements to the XSLTs that convert ODD to PureODD, as well as the resulting new PureODD source, which includes some manual fixes.
The old tei_odds schema has been replaced with a bleeding-edge TEI schema has also been added to properly validate PureODD and references to the schema in driver.xml and mei-source.xml have been updated.

Refs #304