remove unnecessary comment
fixes https://github.com/music-encoding/guidelines/issues/197

the comment in question is too long for one line, and apparently something in the pipeline inserts a line break, which has led to partial re-insertion of the commented code. The result was invalid JS code, which could not be parsed and thus stopped browsers. As the comment is just for debugging, it can safely go away.