xslt for preparing a PDF of the Guidelines
This will generate a file MEI_Guidelines_v5.0.0-dev_latest.html (at the time of writing) next to the mei-source.xml file. With PrinceXML installed, this file needs to be compiled with the command `prince MEI_Guidelines_v5.0.0-dev_latest.html`, which will result in a PDF of the same name and in the same folder.
As of yet, this does not deal with embedding an DOI for non-development compilations.