update schematron rule for mei:history
As `<exhibHist>`, `<treatHist>` and `<treatSched>` have been remodelled after `<provenance>` in #911, the corresponding Schematron rule to report an error in work and expression contexts had to be updated. Moreover, in #934, we agreed also to apply this rule to the manifestation context if the manifestation is not a singleton.
This commit applies the necessary changes.