Fixed: Use value of "schematron" for constraintSpec
The TEI has changed the value of "schematron" in its RNG.

If we're thinking about moving towards pure ODD this change will mean one less thing that breaks.

Also fixed an issue where the customizations were using the TEI RNG, but the sources were pointing to the (older) MEI RNG. These were normalized to the MEI RNG so that all source and customizations use the same ODD schema.

Fixes #1158