build.xml: introduce property `verbose`
Before this commit the generation of schemata and compiled ODDs gave very verbose feedback. This resulted in a very long and hardly graspable output on the commandline. Especially in the GitHub Action context this made the build process hard to debug and required a lot of scrolling.
This commit introduces the following changes:
* The property `verbose` will analyze whether Ant was called with the `-verbose` or `-v` option and store this as boolean; defaults to false.
* The `verbose` property will be propagated to the  `tei:anttasks` to make the verbosity of the schema and compiled-odd generation controllable.