Revise model of respStmt
This PR changes the content model of respStmt to 
- make it more TEI-like ( resp+, nameLike+ or nameLike+, resp+);
- clarify use of respStmt

This is a backward-compatibility breaking change.  However, if the existing model of respStmt has been used to contain only names, each with a role attribute, then that usage can easily be converted via XSLT to the new model; that is, the content of the role attribute can be placed in a new resp element.