Correctly display angle brackets in search queries
When inserting search queries into the title of the search modal, the
code previously set `.innerHTML`, which means that any query that could
be interpreted as HTML (even if poorly formed!) would be rendered. While
this is a markup injection vector, there's no security risk because
search queries aren't sharable (i.e. they aren't stored in the URL).
Instead, this is merely an annoyance to users attempting to search for
XML tags.

Use `.innerText` instead so any HTML-ish search query isn't rendered.